pyproject.toml
[build-system]
requires = ["hatchling", "hatch-fancy-pypi-readme"]
build-backend = "hatchling.build"

[project]
name = "pimoroni-mics6814"
dynamic = ["version", "readme"]
description = "Library for the Pimoroni mics6814 breakout-garden breakout"
license = {file = "LICENSE"}
requires-python = ">= 3.7"
authors = [
    { name = "Philip Howard", email = "phil@pimoroni.com" },
]
maintainers = [
    { name = "Philip Howard", email = "phil@pimoroni.com" },
]
keywords = [
    "Pi",
    "Raspberry",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Hardware",
]
dependencies = [
    "pimoroni-ioexpander>=0.0.5"
]

[project.urls]
GitHub = "https://www.github.com/pimoroni/mics6814-python"
Homepage = "https://www.pimoroni.com"

[tool.hatch.version]
path = "mics6814/__init__.py"

[tool.hatch.build]
include = [
    "mics6814",
    "README.md",
    "CHANGELOG.md",
    "LICENSE"
]

[tool.hatch.build.targets.sdist]
include = [
    "*"
]
exclude = [
    ".*",
    "dist"
]

[tool.hatch.metadata.hooks.fancy-pypi-readme]
content-type = "text/markdown"
fragments = [
  { path = "README.md" },
  { text = "\n" },
  { path = "CHANGELOG.md" }
]

[tool.ruff]
exclude = [
    '.tox',
    '.egg',
    '.git',
    '__pycache__',
    'build',
    'dist'
]
line-length = 200

[tool.codespell]
skip = """
./.tox,\
./.egg,\
./.git,\
./__pycache__,\
./build,\
./dist.\
"""

[tool.isort]
line_length = 200

[tool.check-manifest]
ignore = [
    '.stickler.yml',
    'boilerplate.md',
    'check.sh',
    'install.sh',
    'uninstall.sh',
    'Makefile',
    'tox.ini',
    'tests/*',
    'examples/*',
    '.coveragerc',
    'requirements-dev.txt'
]

[tool.pimoroni]
apt_packages = [""]
configtxt = []
commands = [
    "printf \"Setting up i2c...\n\"",
    "sudo raspi-config nonint do_i2c 0"
]
